requirements.txt
anthropic>=0.25.0
python-dotenv>=1.0.0
PyYAML>=6.0.2
